+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added ipAndPort info in ServerPongEvent @zimuya4153
 - Change ServerNetworkHandler to non-const @zimuya4153
-- Refactor SendPacket event @zimuya4153
+- Refactor SendPacket and ReceivePacket event @zimuya4153
 
 ### Fixed